feat(health): promote Trend view into @repowise-dev/ui
Move the presentational Trend surface (delta metric cards, alert blocks,
single-snapshot empty state, KPI history chart, per-file score-change slope
chart) into a shared TrendView in packages/ui/src/health. It takes data as
props (fetched once at the page level), so no host wiring is needed; web's
trend-tab.tsx re-exports it under the original TrendSection name.
