+++ packages/ui/src/health/index.ts
@@ -39,3 +39,4 @@ export * from "./defect-accuracy-card";
 export * from "./code-health-adapter";
 export * from "./triage-view";
 export * from "./coverage-view";
+export * from "./trend-view";
